requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==24.3.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
dynaconf==3.2.7
fake-useragent==2.0.3
frozenlist==1.5.0
g4f==0.4.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
locate==1.1.1
loguru==0.7.3
lxml==5.3.0
multidict==6.1.0
nest-asyncio==1.6.0
propcache==0.2.1
pyaes==1.6.1
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
win32_setctime==1.2.0
yarl==1.18.3
